+++ scripts/AllFilesExample.py
@@ -59,6 +59,11 @@ for entry in tree:
     # information in the data objects, see the header files in
     # GramsSim/GramsDataObj/include.
 
+    # IF YOU'RE COPYING THIS CODE LINE-FOR-LINE AND LOOP-FOR-LOOP,
+    # STOP! Most of what the code does has no purpose other than to
+    # serve as an example. Look at what the code is doing, and ask if
+    # that's what you want to do.
+
     # Let's pick an arbitrary event. How about run=0, event=3? (Note
     # how by using the prefix "ROOT." we have access to all the C++
     # methods in the GramsDataObj dictionary; in particular, we can
@@ -129,87 +134,3 @@ for entry in tree:
                 if trackID in primaries:
                     print ("ReadoutID", readoutID.Index())
                     break
-                    
-    # The above shows how to "go forward" through the
-    # tracks->hits->clusters->readout chain of information. What
-    # follows is the reverse: How one might back-track up the chain.
-
-    # For every waveform in the event:
-    for ( readoutID, waveform ) in tree.ReadoutWaveforms:
-
-        # Sum the ADC counts in all the bins in the digital
-        # waveform. (There is no scientific reason to do this.)
-        sumADC = sum( waveform.Digital() )
-
-        # If the sum of the ADC counts is greater than some arbitrary
-        # number that I just made up:
-        madeUpNumber = 21969;
-        if sumADC > madeUpNumber:
-
-            # The data objects in the dictionary all of the C++
-            # operator<< defined for them, which simplifies printing
-            # them out in C++. Unfortunately, there doesn't seem to be
-            # a simple way of calling operator<< from within
-            # Python. For these data objects, just display a few
-            # details to prove we have access to their information.
-            print ( "EventID ", tree.EventID.Index(), "ReadoutID", readoutID.Index(), 
-                    "sums to", sumADC, "which is more than", madeUpNumber )
-
-            # Let's assume that this means the waveform is
-            # "interesting". Look at all the electron clusters that went
-            # into creating this waveform.
-            clusterKeys = tree.ReadoutMap[ readoutID ]
-
-            # The cluster keys are, in turn, a set of keys into the
-            # ElectronClusters map. The following looks at all the
-            # electron clusters associated with those keys.
-
-            # If this seems complex, you can think of this arrangement as:
-            #
-            # (a) a multi-dimensional array indexed by [trackID][hitID][clusterID][readoutID];
-            # (b) if you're into Python, nested dictionaries (dicts of dicts).. 
-
-            for clusterKey in clusterKeys:
-                electronCluster = tree.ElectronClusters[ clusterKey ]
-
-                # What can we do with an electron cluster? We might print
-                # it out, according to some imaginary non-scientific
-                # criteria.
-                imaginaryCriteria = 38
-                if electronCluster.NumElectrons() < imaginaryCriteria:
-                    print ("TrackID", electronCluster.TrackID(), "HitID", electronCluster.HitID(), "ClusterID", 
-                           electronCluster.ClusterID(), "energy", electronCluster.EnergyAtAnode())
-
-                    # We can also use it to back-track to a particular
-                    # simulated hit in the LAr. The data object MCLArHits is,
-                    # once again, a map of (key,value) pairs, with the the key
-                    # itself being a pair of (trackID,hitID):
-                    
-                    trackID = electronCluster.TrackID()
-                    hitID   = electronCluster.HitID()
-                    hitKey = ROOT.std.make_tuple( trackID, hitID )
-                    hit = tree.LArHits[ hitKey ]
-
-                    # Now we have the MCLArHit associated with this particular
-                    # electron cluster. Let's apply an arbitrary non-scientific
-                    # cut on the number of scintillation photons to show how to
-                    # work with a hit.
-                    numPhotons = hit.NumPhotons()
-                    unscientificCut = 200
-                    if numPhotons < unscientificCut:
-
-                        # We can print out the hit if we wish.
-                        print ("TrackID", hit.TrackID(), "HitID", hit.HitID(), "energy", hit.Energy())
-
-                        # As a final example of back-tracking, let's go to the
-                        # track that created the hit.
-                        createID = hit.TrackID()
-                        track = tree.TrackList[ createID ]
-
-                        # Now that we have a track, let's look at its trajectory.
-                        trajectory = track.Trajectory()
-
-                        # Let's print the first point in the trajectory:
-                        firstPoint = trajectory[0]
-                        print ( "First point in track (x,y,z,t)", 
-                                firstPoint.X(), firstPoint.Y(), firstPoint.Z(), firstPoint.T() )